snap/restore: check result of SET_MSRS on restore
When restoring a snapshot, Firecracker uses the `KVM_SET_MSRS` ioctl to
restore the values of the Guest MSRs.
Firecracker does not check how many of the requested MSRs were actually
restored by KVM. Since KVM will exit the ioctl early if it can't set a
MSR (see linux/v5.10.143/source/arch/x86/kvm/x86.c#L3693), this may lead
the Guest to potentially missbehave because the values it expects are
not restored correctly.
This issue can be seen if snapshots are restored across different
versions of KVM or different processors that don't have the same
capabilities.
If KVM can't restore all MSRs that are required, we now abort the
restore process and throw and error.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>