snap/restore: save MSRs based on CPUID
Some additional MSRs need to be saved if certain features are enabled in
CPUID.
Add the capability to infer which MSRs to save based on the feature bits
set in CPUID. The dependencies between CPUID features and MSRs are
defined for each CPU vendor.
Through this commit, the list of MSRs we need to save becomes
architectural_msrs + msrs_based_on_cpuid + msrs_defined_in_template.

The dependency list now contains only one entry for the MPX feature bit
so it is by no means exhaustive. We'll add more dependencies in following
commits. Until then the right coverage should be supplied through
templates and the ALLOWED_MSR_RANGES list.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>