snap/restore: extend the number of saved MSRs.
KVM can get/set only a maximum of KVM_MAX_MSR_ENTRIES MSRs at a time
through ioctl.
In some scenarios we need to save more than KVM_MAX_MSR_ENTRIES in the
snapshot when the Guest CPU is emulating a lot of MSRs.

Change the list of supported MSRs to a HashSet. This allows better
flexibility in adding and comparing MSR address lists.
Also save a `Vec<Msrs>` instead of a single `Msrs` struct to allow
saving an unlimited number of MSRs in the snapshot.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>