tests: upgrade snap resume baseline infra
The purpose of this change is to align the test framework with the one
in the main branch (as of 00ebfe26e97258b6f489ab38c47057a38aef7118) to
allow snap restore baseline collection in firecracker-v1.1 and
comparison with the baselines collected in main.
The major difference compensated is the baseline config file schema.

The following has been updated:
 - test_snapshot_restore_performance.py
 - snap restore baseline config files
 - parse_baseline script.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>