fix(tests): test_balloon_snapshot: add delay after inflate
Since the configured stats polling interval is 1 second,
it may take up to 1 second for the `available_memory` stat
to get updated.
In some cases, the VM is snapshotted, restored and
the `available_memory` stat is queried before the 1 second
has passed after an inflation had been requested.
To avoid this situation, a 1 second delay is added after the
inflation request.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>