feat(vm_memory): map guest memory as shared
The change is needed to be able to share the guest memory
with another process that needs access to it
(eg a vhost-user backend).

The change includes opening the memory snapshot file with
read and write permissions as the guest memory changes
due to the guest execution will now be visible
as changes of the memory snapshot file.

A consiquence of that is multiple microVMs can no longer
be restored from the same physical memory snapshot file.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>