feat(vmm): use memfd for guest memory
When sharing guest memory with another process
(eg a vhost-user backend), a file descriptor pointing
at the guest memory is needed that can be passed
to the other process via a sendmsg syscall.
This is achieved via backing guest memory with memfd
when launching a microVM and via making use of
the memory snapshot fd when restoring from a snapshot.

A memfd is sealed after creation to prevent further
changes of the guest memory size (both growing and shrinking).

This change introduces a new crate dependency: memfd.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>