feat(vmm): add vhost-user-blk device
This adds vhost-user block device implementation.

In order to create a vhost-user block device,
a PUT /drives request needs to be issued that
contains `vhost-user` object.
At the moment, the object has the only property `socket`
that is a socket path to communicate with
a vhost-user backend.

Example:
```
curl --unix-socket ${fc_socket} -i \
  -X PUT "http://localhost/drives/vhost" \
  -H "accept: application/json" \
  -H "Content-Type: application/json" \
  -d "{
            \"drive_id\": \"vhost\",
            \"is_root_device\": false,
            \"cache_type\": \"Unsafe\",
            \"vhost_user\": {
              \"socket\": \"${vhost_socket}\"
            }
      }"
```

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>
Co-authored-by: Diana Popa <dpopa@amazon.com>