chore(devctr): build crosvm for a vhost-user-blk backend
This includes a crosvm binary in the devctr.
crosvm is supposed to be used as an additional
vhost-user-blk backend (along with Qemu), because
it has support for drive resizing which we would
like to test at integration level.

We build a fixed crosvm git SHA to be able to
manually control crsovm upgrades.

TODO: images size
TODO: tag

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>