test(vhost-user-blk): add crosvm backend
This adds support for crosvm backend in
spawn_vhost_user_backend() and a new resize()
helper that is only crosvm backend is capable of.

resize() physicly resizes the drive file and
also makes the crosvm backend send a config change
notification to the frontend.

The purpose of the change is to enable us to test
that the frontend (Firecracker) handles config changes
correctly.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>